Extract `replica_set/snapshots.rs`
