+++ lib/collection/src/shards/replica_set/snapshots.rs
@@ -1,0 +1,142 @@
+use super::*;
+
+impl ShardReplicaSet {
+    /// Returns if local shard was recovered from path
+    pub async fn restore_local_replica_from(&self, replica_path: &Path) -> CollectionResult<bool> {
+        if !LocalShard::check_data(replica_path) {
+            return Ok(false);
+        }
+
+        // TODO:
+        //   Check that shard snapshot is compatible with the collection
+        //   (see `VectorsConfig::check_compatible_with_segment_config`)
+
+        let mut local = self.local.write().await;
+
+        // Drop `LocalShard` instance to free resources and clear shard data
+        let clear = local.take().is_some();
+
+        // Try to restore local replica from specified shard snapshot directory
+        let restore = async {
+            if clear {
+                LocalShard::clear(&self.shard_path).await?;
+            }
+
+            LocalShard::move_data(replica_path, &self.shard_path).await?;
+
+            LocalShard::load(
+                self.shard_id,
+                self.collection_id.clone(),
+                &self.shard_path,
+                self.collection_config.clone(),
+                self.shared_storage_config.clone(),
+                self.update_runtime.clone(),
+            )
+            .await
+        };
+
+        match restore.await {
+            Ok(new_local) => {
+                local.replace(Local(new_local));
+                Ok(true)
+            }
+
+            Err(restore_err) => {
+                // Initialize "dummy" replica
+                local.replace(Dummy(DummyShard::new("Failed to restore local replica")));
+
+                // TODO: Handle single-node mode!? (How!? 😰)
+
+                // Mark this peer as "locally disabled"...
+                let has_other_active_peers = self.active_remote_shards().await.is_empty();
+
+                // ...if this peer is *not* the last active replica
+                if has_other_active_peers {
+                    self.locally_disabled_peers
+                        .write()
+                        .insert(self.this_peer_id()); // TODO: Blocking `write` call in async context
+
+                    // Notify peer failure
+                    self.notify_peer_failure_cb.deref()(self.this_peer_id(), self.shard_id);
+                }
+
+                // Remove shard directory, so we don't leave empty directory/corrupted data
+                match tokio::fs::remove_dir_all(&self.shard_path).await {
+                    Ok(()) => Err(restore_err),
+
+                    Err(cleanup_err) => {
+                        log::error!(
+                            "Failed to cleanup shard {} directory ({}) after restore failed: \
+                             {cleanup_err}",
+                            self.shard_id,
+                            self.shard_path.display(),
+                        );
+
+                        // TODO: Contextualize `restore_err` with `cleanup_err` details!?
+                        Err(restore_err)
+                    }
+                }
+            }
+        }
+    }
+
+    pub fn restore_snapshot(
+        snapshot_path: &Path,
+        this_peer_id: PeerId,
+        is_distributed: bool,
+    ) -> CollectionResult<()> {
+        let replica_state: SaveOnDisk<ReplicaSetState> =
+            SaveOnDisk::load_or_init(snapshot_path.join(REPLICA_STATE_FILE))?;
+
+        // If this shard have local data
+        let is_snapshot_local = replica_state.read().is_local;
+
+        if !is_distributed && !is_snapshot_local {
+            return Err(CollectionError::service_error(format!(
+                "Can't restore snapshot is local mode with missing data at shard: {}",
+                snapshot_path.display()
+            )));
+        }
+
+        replica_state.write(|state| {
+            state.this_peer_id = this_peer_id;
+            if is_distributed {
+                state
+                    .peers
+                    .remove(&this_peer_id)
+                    .and_then(|replica_state| state.peers.insert(this_peer_id, replica_state));
+            } else {
+                // In local mode we don't want any remote peers
+                state.peers.clear();
+                state.peers.insert(this_peer_id, ReplicaState::Active);
+            }
+        })?;
+
+        if replica_state.read().is_local {
+            LocalShard::restore_snapshot(snapshot_path)?;
+        }
+        Ok(())
+    }
+
+    pub async fn create_snapshot(
+        &self,
+        temp_path: &Path,
+        target_path: &Path,
+        save_wal: bool,
+    ) -> CollectionResult<()> {
+        let local_read = self.local.read().await;
+
+        if let Some(local) = &*local_read {
+            local
+                .create_snapshot(temp_path, target_path, save_wal)
+                .await?
+        }
+
+        self.replica_state
+            .save_to(target_path.join(REPLICA_STATE_FILE))?;
+
+        let shard_config = ShardConfig::new_replica_set();
+        shard_config.save(target_path)?;
+        Ok(())
+    }
+}
